requirements.txt
Django==1.11.17
django-cors-headers==3.1.0
gunicorn==19.9.0
djangorestframework==3.10.2
